Add custom clone path for the build
When the user specifies `CI_PROJECT_DIR` environment variable inside of
the `.gitlab-ci.yml` if the path is valid it is used to clone and build.

`CI_PROJECT_DIR` takes precedence over the configured build dir and
the default build dir.

This is not available when shared dir are enabled meaning:
- Docker executor has a mounted volume
- Shell executor
- SSH executor

Closes #2211