Start build container detached with tty
The way the build script is executing on the build container is by
starting the container with the script as ps1, as soon as the script is
done, the container exists.

Follow the same flow the kubernetes executor, where the build container
is started in detached mode and the build script are executed on top of
the running build container.

This is done so interactive web terminal have the same behavior as
kubernetes, where even after the main script is done the user stays
connected to the web terminal.