Add windows Dockerfiles for gitlab-runner-helper
Add the initial Dockerfiles for windows server, using servercore as the
bilder image, which gives us access to powershell, and nanoserver as the
final image to keep the image as small as possible.

These Dockerfiles expect to have
`./binaries/gitlab-runner-helper.win_amd64.exe` built already, it will
NOT build the binary itself to stay consistent with the Linux
Dockerfiles and reduce duplication.

We need to add a Dockerfile for each windows server version that we want
to support, because when you run windows containers the host machine and
the container version has to match. For example of your machine that
has docker installed in running `ltsc2019` the containers it can run are
only of the version `ltsc2019`