Add script for building/releasing windows image
The decision for using PowerShell for windows is because we cannot
build/save/export windows containers to a non-Windows image so we to use
a scripting language native to windows. Bash can be used but it would be
in a subsystem which is not ideal. This script will be used inside of
GitLab CI in a future MR when the infrastructure has set up
https://gitlab.com/gitlab-com/gl-infra/infrastructure/issues/5952.

This PowerShell script is doing a few things different than
`release_docker_images` for the helper images. In the bash script we
expect the helper images are already built and exported using `docker
export` but for windows containers when you run the command you get `the
daemon on this operating system does not support exporting Windows
containers`. This script will build the docker container and then push
it to the docker registry hub.