Fix tag for linux helper image locally
With https://gitlab.com/gitlab-org/gitlab-runner/merge_requests/1174 we
changed the tags for the linux based helper image to
`gitla/gitlab-runner-helper:linux_{{$ARCH}}-{{$HASH}}` which is not what
we expect when pulling the image down, we expect
`gitla/gitlab-runner-helper:linux_{{$ARCH}}`. This commit reverts back
to the original tag as expected.