Don't use SIGKILL on su process of shell executor
Using SIGKILL on the su process used to spawn a shell job just kills the
process and prevents it from telling anybody about it's untimely demise.
It's basically MIA and it's only when people start asking where s/he is
that we find out he's been disappeared as a political dissident.

Instead, send it SIGTERM so that our actual job can be stopped and torn
down in an orderly (and complete) fashion.  In its current state,
canceling a job leaves my entire build job running (both on the host and
the docker container I spawn from my build script).

This patch may not be perfect.  I've never touched golang before and I
have no clue why the pid is sometimes stored as a negative number, but
it's far better than it was before.  Even sending INT, HUP or TERM
followed by a delay of 50ms before sending KILL should be enough for the
su process to send HUP to its kids in most cases (depending upon the
process scheduler and load of course).  The most ideal solution would be
to send an INT or TERM and join on it for 1000ms.  If it still hasn't
gone away, go ahead and send KILL, but never send it as the first
signal!