Add feature flag to use new helper image cmd
With https://gitlab.com/gitlab-org/gitlab-runner/merge_requests/1195 &
https://gitlab.com/gitlab-org/gitlab-runner/merge_requests/1201 the bash
scripts were moved to Go so that they can be used for windows
containers.

If we change the command it would lead into a breaking change for the
users to specify which helper image to use.

Introduce a new FF that will allow linux users to use the new command.