Add config option for windows helper image
Windows containers have to be the same windows server version as the
dockerd host is running. Meaning is dockerd is running on windows server
1809, all the images that is runs has to depend on 1809.

Having a configuration specific for windows where the user has to
specify which version the user wants to use is the simplest solution.
Since it's not a simple task specially when Dokcer is running on a
remote host.

https://docs.docker.com/engine/api/v1.39/#operation/SystemInfo might
have the information we need for future iteration.