Refactor helpers/docker/helperimage package
The helper image is being used for both Kubernetes and Docker executor.
But the code does not represent this, instead it duplicates the logic,
such as creating the Linux tag and the name of the image.

Refactor the helperimage package so that it can be used for both, by
removing specific structs from the docker API, by defining helperimage
own Config struct.

The need for the refactor is for
https://gitlab.com/gitlab-org/gitlab-runner/merge_requests/1243, because
we need to define different commands to run on the helper image, which
is located in `common/container.go` which needs to change depending on
the OS Type of docker, and we can't change it before we change how
Kubernetes uses the helper image.