Fix gofmt errors
We use codeclimate to run checks on our source code, and gofmt which is
based on https://github.com/codeclimate-community/codeclimate-gofmt.
This was using go 1.9 `gofmt` but was updated recently to 1.12 in
https://github.com/codeclimate-community/codeclimate-gofmt/commit/098a4f8af0da7770b0e45311dec33770da0e2552
which has updated gofmt rules.

Version pinning to an old version of gofmt is not possible since
codeclimate only provide 1 channel
https://github.com/codeclimate/codeclimate/blob/eab15e5f89704f7ff63f66bfd33d03f060f6c3c8/config/engines.yml#L82-L84