Print correlation_id for GitLab requests
Log the `X-Request-Id` better known as `correlation_id` when we get a
response from GitLab coordinator. This will allow administrators use the
correlation_id and use distributed tracing to find out the strack trace
in rails, and other applications involved. More information can be found
in
https://docs.gitlab.com/ee/development/distributed_tracing.html#using-jaeger-in-the-gitlab-development-kit,
this can also be used to filter in sentry.