Add support for Windows Server 2016
This adds support for Windows Server 2016, build 1607. For the most part
the setup is the same as with newer versions of Windows (Server), with
one difference: the ContainerAdministrator user is not explicitly
specified. When setting this user in the helper Dockerfile, the build
will fail with an error along the lines of "a password is required".
Based on a variety of Docker issues I looked at, this seems to be some
sort of limitation of Windows Server 2016. Fortunately, removing the
USER line solves the problem and the resulting image works just fine.