Scrub URLs from log keys
SecretsCleanupHook removes secrets from log messages such as
`X-AMZ-Signature`. It does not remove them for the keys, so when a URL
with tokens are logged they are not scrubbed as expected.

Iterate over the keys and scrub them as as well, since this is quite a
hot path (every log uses this hook) a benchmark was added to test
performance to calculate the before and after and there is a 3x
performance hit. After an attempt of adding concurrency the performance
was even worse since a mutex had to be introduced.

pkg: gitlab.com/gitlab-org/gitlab-runner/log
BenchmarkSecretsCleanupHook_Fire-8   	 1000000	      1326 ns/op
PASS

pkg: gitlab.com/gitlab-org/gitlab-runner/log
BenchmarkSecretsCleanupHook_Fire-8   	  300000	      4813 ns/op
PASS