Update mocks command to run on all helper pkg
We have a lot of packages under the `helpers` package and we have some
mocks not being generated becuase they are not added to the list.
Instead of having to remember to add one package at at time just use the
helepr dir.

This was a real issue for the interfaces defined in `helpers/tls`