Refactor metrics referee test
Add some spacing and remove the unnecessary cancel context since we
aren't really using it, an empty `context.Backgound()` is better in this
case.

Use `assert` instead of `require` in most places in case of failure so
that the tests run all the assertions to give as much information
possible to the developer.

Rename some tests to be more clear what they are testing.