Refactor referee life cycle
Leave it up the Build to control the life cycle of the referee instead
of having a split situation between multi and Build. This makes the API
a lot simpler and all the execution/setting up of referees in 1 place.

This will only set up the referees after the build has finished put the
most important thing at the top (the build itself), any failure on the
referee doesn't contribute the build status since the build has already
finished.

Since everything is inside of the Build, we don't need to have an
interface for the provider to check if it implements that interface, when
we are checking if we need to set up a referee.

There is 1 failing test since I ran out of time
`TestMetricsRefereeExecute` which the set up already done, and some
comments on what needs to be done to finish it.