Refactor order of methods for prepareHostAlias
We try to keep a consistent order of the codebase, where a method is put
after the caller of the method.

In this case we have `setupBuildPod` which calls to methods
`prepareHostAlias`, `preparePodConfig` and `makePodProxyServices` so the
order should be like the following:

- `setupBuildPod`
    - `prepareHostAlias`
       - `createHostAlias`
    - `preparePodConfig`
        - Any other method that `preparePodConfig` calls
    - `makePodProxyServices`

Think of it as reading a book where you have 1 entry point and then you
can follow the trail of the full method.