Use vendor dir for test/builds
We have the `vendor` directory so that we are aware of how the Runner is
being build. If the project is cloned outside of the `$GOPATH` the, any
`go` cmd will pull everything from the internet unless you specify to
use the `vendor` dir.

This shouldn't be the case in 1.14
https://github.com/golang/go/issues/33848 anymore but at the moment we
are always pulling everything from the internet for building for example
https://gitlab.com/gitlab-org/gitlab-runner/-/jobs/444793271.

reference https://gitlab.com/gitlab-org/gitlab-runner/issues/5045