Log warn message when networksManager is not defined
When the networksManager is not defined the clean up will return an
`errNetworksManagerUndefined` in
https://gitlab.com/gitlab-org/gitlab-runner/-/blob/master/executors/docker/docker.go#L630-632.
This will check the error type and it will not log it as an Error but a
Warning.

This can be a common case where `Cleanup` is called in the middle of
`Prepare` execution, for example when the user cancels the build. We saw
such case in
https://gitlab.com/gitlab-org/gitlab-runner/-/issues/22485#note_303769660.

Tests are a bit hard to add since we don't have the correct logging
setup for the Builder logger and we can't mock it. This could greatly
improved in https://gitlab.com/gitlab-org/gitlab-runner/-/issues/6609

closes https://gitlab.com/gitlab-org/gitlab-runner/issues/22485