Fix flaky test `TestDockerCommandRunAttempts`
Background:
When
[removeContaienr](https://gitlab.com/gitlab-org/gitlab-runner/-/blob/45d80690755a893d9d8986f08e94778dfe811031/executors/docker/docker_command_test.go#L1289)
is called the exit code is 137. Which is sometimes picked up by the
Runner and that exit code is returned as explained in detail in
https://gitlab.com/gitlab-org/gitlab-runner/-/issues/25385#note_324486793
this causes the test to fail and lead into flaky tests.

Fix:
Add testAttempts, which will run the tests x amount of times, increasing
the odds of us getting the expected value. If the test attempts are
higher there is a high chance that the failure is legit.

The Runner can just retry the job section if the exit code is 137, but
it's useful to show the `exit code 137` to the user because this is
mostly the oom killer killing the container as explained in
https://success.docker.com/article/what-causes-a-container-to-exit-with-code-137
and it's useful to show to the user that their job is getting killed by
the oom killer.

Reproduce:
It is quite hard to reproduce since it's random and no real way to
provide the exit code properly. The only way that you can is calling
`ContainerStop` with the following patch:
https://gitlab.com/snippets/1967002

closes https://gitlab.com/gitlab-org/gitlab-runner/-/issues/25385