Remove prepare done job for DAG
The `prepare done` job was added so that we can have jobs start as soon
as the pipeline is created, since there wasn't a point to wait for the
previous stage to finish. With `needs` we can specify `needs: []` which
does the same thing it will start the job imideally and doesn't wait for
the previous stage as shown in
https://docs.gitlab.com/ee/ci/yaml/#needs.

With this change we also solve a problem for `docs/` only merge requests
becuase `prepare done` wasn't running on those, which `check feature
flags` (which runs on docs/ branches) required.

closes https://gitlab.com/gitlab-org/gitlab-runner/-/issues/26593