Use Go Cloud for S3 transfers
The current mechanism of using pre-signed URLs is limited to 5 GB and
slower than using a parallel, multipart upload mechanism that is
implemented in the AWS SDK.

When static credentials are used, this commit exports these credentials
to the cache-archiver as environment variables for the build stages that
upload or download from the cache. IAM instance profile credentials are
retrieved by the AWS SDK inside the cache-archiver.