powershell: Fix commands no longer executing
https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/3831
reverted the change for passing parameters via stdin with
`-EncodedCommand` with `powershell` but kept this for `pwsh`.
However, there appear to be some issues using `-Command` in some
cases. Jobs can fail with:

```
echo : The term '﻿echo' is not recognized as the name of a cmdlet,
function, script file, or operable program.  Check the spelling of the
name, or if a path was included, verify that the path is correct and
try again.
At line:1 char:1
```

It appears that including the UTF-8 BOM may be causing problems with
`-Command`, and thus we should revert to the previous behavior of
passing the command via file.

Relates to https://gitlab.com/gitlab-org/gitlab-runner/-/issues/29551