Log images used by job
Depending on the used executor, a job may be configured
with an image and/or service images to use.

In some cases it will be very useful to the Runner owner to track what
images are used: for copliance reasons, for statistics, to know what to
pre-provision in the job environment or to check whether there are
images in use that may break the jobs once a specific configuration
change to the runner is applied.

This commit adds logging of these images.

Depending on the specific configuration of Runner and most importantly -
the job - this change may significantly increase the number of log
output produced by Runner. Therefor, for now, it's hidden behind a
feature flag that is disabled by default.