Fix docker image option types
This is the type hierarchy required to add `image.docker.platform`
options to both job and service images. A dedicated type is neater than
a `map[string]any` as we a) get type checking on unmarshalling for free,
and b) it makes for a nicer API to pull out the values.