Method to send SIGTERM to all processes in container
Running containers with a proper init system is not enough to guarantee
graceful shutdown of the container because shells effectively swallow
signals instead of propagating them to child processes.

As an alternate, more general and robust approach, this method will exec
into the specified container and run the script introduced earlier,
which sends SIGTERM to all processes in the container in descending
numeric PID order.